+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `app download --outdated` flag to download new versions of all outdated apps
+
+### Changed
+
+- Removed `json` flag from `app download` command
+
 ## [1.15.0] - 2024-03-11
 
 ### Fixed